pyproject.toml
[project]
name = "youtu-agent"
version = "0.1.3"
description = "Youtu-Agent is a simple but powerful framework for building, running, and evaluating agents."
authors = [{ name = "easonsshi", email = "easonsshi@tencent.com" }]
license = "MIT"
keywords = ["agent", "ai", "llm", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.15",
    "art>=6.5",
    "colorlog>=6.9.0",
    "hydra-core>=1.3.2",
    "jinja2>=3.1.6",
    "mcp>=1.12.3",
    "openai-agents==0.10.4",
    # https://github.com/Arize-ai/openinference/tree/main/python/instrumentation/openinference-instrumentation-openai
    "openinference-instrumentation-openai>=0.1.41",
    # https://github.com/Arize-ai/openinference/tree/main/python/instrumentation/openinference-instrumentation-openai-agents
    "openinference-instrumentation-openai-agents>=1.4.0",
    # https://opentelemetry.io/docs/languages/python/
    "opentelemetry-exporter-otlp>=1.37.0",
    "prompt-toolkit>=3.0.52",
    "python-frontmatter>=1.1.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "sqlmodel>=0.0.24",
]

[project.urls]
Homepage = "https://github.com/TencentCloudADP/youtu-agent"
Documentation = "https://tencentcloudadp.github.io/youtu-agent/"
Repository = "https://github.com/TencentCloudADP/youtu-agent"

[project.optional-dependencies]
litllm = ["litellm>=1.77.0"]
search = ["crawl4ai>=0.7.4"]
scholar = ["arxiv>=2.3.0"]
wiki = ["wikipedia-api>=0.8.1"]
e2b = ["e2b>=2.5.0", "e2b-code-interpreter>=2.2.1"]
local-python = [
    "chardet>=5.2.0",
    "ipython>=8.37.0",
    "matplotlib>=3.10.6",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
]
local-bash = ["pexpect>=4.9.0"]
documents = [
    "chunkr-ai>=0.3.7",
    "openpyxl>=3.1.5",
    "pymupdf>=1.26.4",
    "python-magic>=0.4.27",
    "python-pptx>=1.0.2",
    "unstructured[all-docs]>=0.18.14",
]
image = ["pillow>=11.3.0"]
video = ["google-genai>=1.38.0"]
utils = ["docker>=7.1.0", "tiktoken>=0.11.0"]
# PPT generation example dependencies
ppt-gen = [
    "pillow>=11.3.0",
    "python-pptx>=1.0.2",
    "pymupdf>=1.26.4",
    "matplotlib>=3.10.6",
]


[dependency-groups]
dev = [
    "arize-phoenix-client>=1.14.1",
    "arize-phoenix-otel>=0.11.0",
    # datasets == 4.0.0 remove scripts, https://github.com/huggingface/datasets/issues/7693
    "datasets==3.6.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.16",
    "mkdocstrings[python]>=0.30.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.8",
    "tornado>=6.5.2",
    "twine>=6.2.0",
]
gaia = [
    "astor>=0.8.1",
    "docx2markdown>=0.1.1",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "pymupdf>=1.26.3",
    "retry>=0.9.2",
    "tabulate>=0.9.0",
    "unstructured[docx,pptx]>=0.18.14",
    "validators>=0.35.0",
    "xls2xlsx>=0.2.0",
    "xmltodict>=0.15.0",
]
rl = ["aiolimiter>=1.2.1", "tencentcloud-sdk-python>=3.0.1474"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utu"]
# exclude = ["utu/ui"]
# only-include = ["utu/utils"]


[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = ["-v", "-s"]

[tool.uv]
index-url = "https://pypi.org/simple"
# extra-index-url = [
#     "https://mirrors.tencent.com/repository/pypi/tencent_pypi/simple",
# ]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = ["**/*.pyi", ".*", "_*.py", "configs/", "utu/patch/runner.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
isort = { combine-as-imports = true, known-first-party = ["utu"] }

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.mypy]
strict = true
disallow_incomplete_defs = false
disallow_untyped_defs = false
disallow_untyped_calls = false

# [tool.pylint.messages_control]
# disable = ["broad-except", "bare-except", "invalid-name"]
